Re Wrote assets managers
