+++ constants.py
@@ -1,0 +1,10 @@
+import os
+
+'''
+    All Recurring Constants go in this file. Do not Forget to invoke "import constants" in any python file you need this
+
+'''
+
+# Directory Related constants
+script_dir = os.path.dirname(__file__)
+jar_dir = "\\jar"
