+++ recipeviewer.py
@@ -1,0 +1,4 @@
+import PySimpleGUI as sg 
+import constants
+
+print(constants.script_dir)
